+++ R/learnr-package.R
@@ -1,0 +1,34 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom evaluate evaluate
+#' @importFrom htmltools attachDependencies
+#' @importFrom htmltools div
+#' @importFrom htmltools HTML
+#' @importFrom htmltools htmlDependency
+#' @importFrom htmltools tags
+#' @importFrom htmlwidgets createWidget
+#' @importFrom jsonlite base64_dec
+#' @importFrom jsonlite base64_enc
+#' @importFrom knitr all_labels
+#' @importFrom knitr knit_hooks
+#' @importFrom knitr knit_meta_add
+#' @importFrom knitr opts_chunk
+#' @importFrom knitr opts_hooks
+#' @importFrom knitr opts_knit
+#' @importFrom knitr spin
+#' @importFrom markdown markdownExtensions
+#' @importFrom markdown markdownToHTML
+#' @importFrom rprojroot find_root
+#' @importFrom rprojroot is_r_package
+#' @importFrom shiny invalidateLater
+#' @importFrom shiny isolate
+#' @importFrom shiny observe
+#' @importFrom shiny observeEvent
+#' @importFrom shiny reactive
+#' @importFrom shiny reactiveValues
+#' @importFrom shiny req
+#' @importFrom withr with_envvar
+## usethis namespace: end
+NULL